HAWQ-1100. Handle as double not float

+++ pxf/pxf-hbase/src/main/java/org/apache/hawq/pxf/plugins/hbase/utilities/HBaseDoubleComparator.java
@@ -60,7 +60,7 @@ public class HBaseDoubleComparator extends ByteArrayComparable{
             throw new DeserializationException(e);
         }
 
-        return new HBaseDoubleComparator(Bytes.toFloat(proto.getValue().toByteArray()));
+        return new HBaseDoubleComparator(Bytes.toDouble(proto.getValue().toByteArray()));
     }
 }
